python-backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12
PyPDF2==3.0.1
pdfplumber==0.11.4
python-dotenv==1.0.1
supabase==2.9.0
numpy==1.26.4
scikit-learn==1.5.1
structlog==24.2.0
PyYAML==6.0.2
rapidfuzz==3.9.3
rank-bm25==0.2.2
regex==2024.9.11
